# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dealit API
  description: Dealit backend API documentation
  contact:
    name: Dealit Team
  version: v1
servers:
- url: http://localhost:8080
  description: Local server
security:
- bearerAuth: []
...
paths:
...
  "/api/v1/users/me/profile-image":
    post:
      tags:
      - Profile
      summary: 프로필 이미지 업로드
      description: 현재 로그인한 사용자의 프로필 이미지를 업로드합니다.
      operationId: uploadProfileImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UploadProfileImageResponse"
...
  "/api/v1/products/price/recommend":
    post:
      tags:
      - Product
      summary: 가격 추천
      operationId: recommendPrice
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RecommendPriceRequest"
        required: true
      responses:
        '200':
          description: 가격 추천 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RecommendPriceResponse"
  "/api/v1/products/image":
    post:
      tags:
      - Product
      summary: 상품 이미지 업로드
      operationId: uploadImage
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        '200':
          description: 이미지 업로드 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UploadProductImageResponse"
  "/api/v1/products/draft":
    post:
      tags:
      - Product
      summary: 일반 상품 임시저장
      operationId: saveDraft
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SaveProductDraftRequest"
        required: true
      responses:
        '200':
          description: 임시저장 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SaveProductDraftResponse"
  "/api/v1/products/category/recommend":
    post:
      tags:
      - Product
      summary: 카테고리 추천
      operationId: recommendCategory
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RecommendCategoryRequest"
        required: true
      responses:
        '200':
          description: 카테고리 추천 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RecommendCategoryResponse"
  "/api/v1/members/signup":
    post:
      tags:
      - Member
      summary: 회원가입
      description: 신규 회원을 생성합니다.
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SignUpRequest"
        required: true
      responses:
        '201':
          description: Created
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SignUpResponse"
...
  "/api/v1/email/verification/send":
    post:
      tags:
      - Email Verification
      summary: 이메일 인증 코드 발송
      description: 회원가입 또는 프로필 수정에 사용할 이메일 인증 코드를 발송합니다.
      operationId: send
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SendEmailVerificationRequest"
        required: true
      responses:
        '201':
          description: Created
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SendEmailVerificationResponse"
  "/api/v1/email/verification/confirm":
    post:
      tags:
      - Email Verification
      summary: 이메일 인증 코드 확인
      description: 발송된 이메일 인증 코드를 검증합니다.
      operationId: confirm
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ConfirmEmailVerificationRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ConfirmEmailVerificationResponse"
...
  "/api/v1/auction":
    post:
      tags:
      - Auction
      summary: 경매 상품 등록
      description: 프론트가 가공한 최종 payload로 상품을 등록한다.
      operationId: createAuction
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateAuctionRequest"
        required: true
      responses:
        '201':
          description: 상품 등록 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CreateAuctionResponse"
  "/api/v1/auction/price/recommend":
    post:
      tags:
      - Auction
      summary: 가격 추천
      description: 상품 정보와 판매 유형을 기반으로 추천 가격을 반환한다.
      operationId: recommendPrice_1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RecommendPriceRequest"
        required: true
      responses:
        '200':
          description: 가격 추천 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RecommendPriceResponse"
  "/api/v1/auction/image":
    post:
      tags:
      - Auction
      summary: 상품 이미지 업로드
      description: 경매 상품 등록 전 이미지를 업로드하고 임시 URL을 반환한다.
      operationId: uploadImage_1
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        '200':
          description: 이미지 업로드 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UploadAuctionImageResponse"
  "/api/v1/auction/draft":
    post:
      tags:
      - Auction
      summary: 경매 상품 임시저장
      description: 등록 직전 폼 데이터를 draft 상태로 저장한다.
      operationId: saveDraft_1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SaveAuctionDraftRequest"
        required: true
      responses:
        '200':
          description: 임시저장 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/SaveAuctionDraftResponse"
  "/api/v1/auction/category/recommend":
    post:
      tags:
      - Auction
      summary: 카테고리 추천
      description: 입력된 상품명과 설명을 기준으로 추천 카테고리를 반환한다.
      operationId: recommendCategory_1
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RecommendCategoryRequest"
        required: true
      responses:
        '200':
          description: 카테고리 추천 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RecommendCategoryResponse"
  "/api/v1/users/me/profile":
    patch:
      tags:
      - Profile
      summary: 프로필 수정
      description: 현재 로그인한 사용자의 닉네임, 소개글, 프로필 이미지를 수정합니다.
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateMyProfileRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MyPageProfileResponse"
  "/api/v1/users/me/location":
    get:
      tags:
      - Profile
      summary: 내 지역 조회
      description: 현재 로그인한 사용자의 지역 및 구조화된 위치 정보를 조회합니다.
      operationId: getMyLocation
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MyLocationResponse"
    patch:
      tags:
      - Profile
      summary: 지역 수정
      description: 현재 로그인한 사용자의 대표 지역과 상세 위치 정보를 수정합니다.
      operationId: updateLocation
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateMyLocationRequest"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/UpdateMyLocationResponse"
...
  "/api/v1/members/nickname/check":
    get:
      tags:
      - Member
      summary: 닉네임 중복 확인
      description: 프로필 설정 화면에서 닉네임 사용 가능 여부를 확인합니다.
      operationId: checkNickname
      parameters:
      - name: nickname
        in: query
        required: true
        schema:
          type: string
          maxLength: 30
          minLength: 2
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/NicknameCheckResponse"
  "/api/v1/members/login-id/check":
    get:
      tags:
      - Member
      summary: 로그인 아이디 중복 확인
      description: 회원가입 화면에서 로그인 아이디 사용 가능 여부를 확인합니다.
      operationId: checkLoginId
      parameters:
      - name: loginId
        in: query
        required: true
        schema:
          type: string
          maxLength: 30
          minLength: 0
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/LoginIdCheckResponse"
...
  "/api/v1/products/image/{imageId}":
    delete:
      tags:
      - Product
      summary: 상품 이미지 삭제
      operationId: deleteImage
      parameters:
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 이미지 삭제 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DeleteProductImageResponse"
  "/api/v1/auction/image/{imageId}":
    delete:
      tags:
      - Auction
      summary: 상품 이미지 삭제
      description: 등록 전에 업로드된 임시 이미지를 imageId 기준으로 삭제한다.
      operationId: deleteImage_1
      parameters:
      - name: imageId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: 이미지 삭제 성공
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DeleteAuctionImageResponse"
components:
  schemas:
...
    UploadProfileImageResponse:
      type: object
      properties:
        profileImageUrl:
          type: string
...
    RecommendPriceRequest:
      type: object
      description: 일반 상품 가격 추천 요청
      properties:
        name:
          type: string
          description: 상품명
          example: 닌텐도 스위치 OLED
          maxLength: 100
          minLength: 0
        description:
          type: string
          description: 상품 설명
          example: 사용하지 않아서 깨끗합니다.
          maxLength: 2000
          minLength: 0
        saleType:
          type: string
          description: 판매 유형
          enum:
          - REGULAR
          - AUCTION
          example: REGULAR
      required:
      - description
      - name
      - saleType
    RecommendPriceResponse:
      type: object
      description: 가격 추천 응답
      properties:
        recommendedPrice:
          type: number
          description: 추천 판매가
          example: 250000
        startPrice:
          type: number
          description: 추천 시작가
          example: 180000
    UploadProductImageResponse:
      type: object
      description: 일반 상품 이미지 업로드 응답
      properties:
        imageId:
          type: integer
          format: int64
          description: 업로드된 이미지 ID
          example: 301
        imageUrl:
          type: string
          description: 업로드된 이미지 URL
          example: http://localhost:8080/product/images/301-sample.jpg
    SaveProductDraftRequest:
      type: object
      description: 일반 상품 임시저장 요청
      properties:
        name:
          type: string
          description: 상품명
          example: 일반 판매 상품명
          maxLength: 100
          minLength: 0
        description:
          type: string
          description: 상품 설명
          example: 상품 설명
          maxLength: 2000
          minLength: 0
        saleType:
          type: string
          description: 판매 유형
          enum:
          - REGULAR
          - AUCTION
          example: REGULAR
        categoryId:
          type: integer
          format: int64
          description: 카테고리 ID
          example: 19
        price:
          type: number
          description: 일반 판매가
          example: 12000
        allowOffer:
          type: boolean
          description: 가격 제안 허용 여부
          example: false
        images:
          type: array
          description: 상품 이미지 목록
          items:
            "$ref": "#/components/schemas/ProductImagePayload"
        location:
          type: string
          description: 거래 위치
          example: 서울 강남구
          maxLength: 100
          minLength: 0
        draftId:
          type: integer
          format: int64
          description: 임시저장 ID
    SaveProductDraftResponse:
      type: object
      description: 일반 상품 임시저장 응답
      properties:
        draftId:
          type: integer
          format: int64
          description: 임시저장 ID
          example: 10
        savedAt:
          type: string
          format: date-time
          description: 저장 시각
          example: 2026-04-12 20:00:00.000000000 Z
    RecommendCategoryRequest:
      type: object
      description: 일반 상품 카테고리 추천 요청
      properties:
        name:
          type: string
          description: 상품명
          example: 닌텐도 스위치
          maxLength: 100
          minLength: 0
        description:
          type: string
          description: 상품 설명
          example: 사용하지 않아서 깨끗합니다.
          maxLength: 2000
          minLength: 0
      required:
      - description
      - name
    RecommendCategoryResponse:
      type: object
      description: 카테고리 추천 응답
      properties:
        categoryId:
          type: integer
          format: int64
          description: 추천 카테고리 ID
          example: 200
        categoryName:
          type: string
          description: 추천 카테고리명
          example: 디지털/가전
    SignUpRequest:
      type: object
      description: 회원가입 요청
      properties:
        loginId:
          type: string
          description: 로그인 아이디
          example: dealit_user
          maxLength: 30
          minLength: 0
        password:
          type: string
          description: 비밀번호
          example: Password123!
          maxLength: 30
          minLength: 8
        email:
          type: string
          format: email
          description: 이메일
          example: hong@example.com
          maxLength: 100
          minLength: 0
        name:
          type: string
          description: 이름
          example: 홍길동
          maxLength: 30
          minLength: 0
      required:
      - loginId
      - password
    SignUpResponse:
      type: object
      description: 회원가입 응답
      properties:
        memberId:
          type: integer
          format: int64
          description: 회원 ID
          example: 1
        loginId:
          type: string
          description: 로그인 아이디
          example: delit_user
        email:
          type: string
          description: 이메일
          example: hong@example.com
        nickname:
          type: string
          description: 자동 생성 닉네임
          example: 당근한입
        createdAt:
          type: string
          format: date-time
          description: 생성 시각
...
    SendEmailVerificationRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 100
          minLength: 0
      required:
      - email
    SendEmailVerificationResponse:
      type: object
      properties:
        email:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
    ConfirmEmailVerificationRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 100
          minLength: 0
        code:
          type: string
          minLength: 1
          pattern: "\\d{6}"
      required:
      - code
      - email
    ConfirmEmailVerificationResponse:
      type: object
      properties:
        email:
          type: string
        verified:
          type: boolean
...
    CreateAuctionRequest:
      type: object
      description: 경매 상품 등록 요청
      properties:
        name:
          type: string
          description: 상품명
          example: 맥북 에어 M1
          maxLength: 100
          minLength: 0
        description:
          type: string
          description: 상품 설명
          example: 얼마 안 사용해서 깨끗합니다.
          maxLength: 2000
          minLength: 0
        saleType:
          type: string
          description: 판매 유형
          enum:
          - AUCTION
          example: AUCTION
        categoryId:
          type: integer
          format: int64
          description: 카테고리 ID
          example: 101
        price:
          type: number
          description: 일반 판매가
          example: 120000
        startPrice:
          type: number
          description: 경매 시작가
          example: 80000
        auctionDurationDays:
          type: integer
          format: int32
          description: 경매 진행 기간(일 단위)
          example: 3
        images:
          type: array
          items:
            "$ref": "#/components/schemas/ProductImagePayload"
            minItems: 1
          minItems: 1
        location:
          type: string
          description: 거래 위치
          example: 서울 강남구
          maxLength: 100
          minLength: 0
        draftId:
          type: integer
          format: int64
          description: 임시저장 ID
          example: 5
      required:
      - categoryId
      - description
      - images
      - location
      - name
      - saleType
...
    CreateAuctionResponse:
      type: object
      description: 경매 상품 등록 응답
      properties:
        productId:
          type: integer
          format: int64
          description: 등록된 상품 ID
          example: 1001
        saleType:
          type: string
          description: 판매 유형
          enum:
          - AUCTION
          example: AUCTION
        status:
          type: string
          description: 상품 상태
          enum:
          - DRAFT
          - AUCTION_LIVE
          - ENDED
          example: AUCTION_LIVE
        auction:
          "$ref": "#/components/schemas/AuctionScheduleResponse"
          description: 경매 메타 정보
    UploadAuctionImageResponse:
      type: object
      description: 상품 이미지 업로드 응답
      properties:
        imageId:
          type: integer
          format: int64
          description: 업로드된 이미지 ID
          example: 301
        imageUrl:
          type: string
          description: 업로드된 이미지 URL
          example: http://localhost:8080/auction/images/301-sample.jpg
    SaveAuctionDraftRequest:
      type: object
      description: 경매 상품 임시저장 요청
      properties:
        name:
          type: string
          description: 상품명
          example: 애플워치 시리즈 9
          maxLength: 100
          minLength: 0
        description:
          type: string
          description: 상품 설명
          example: 개봉 후 몇 번 사용하지 않았고 충전기도 함께 드립니다.
          maxLength: 2000
          minLength: 0
        saleType:
          type: string
          description: 판매 유형
          enum:
          - REGULAR
          - AUCTION
          example: AUCTION
        categoryId:
          type: integer
          format: int64
          description: 카테고리 ID
          example: 101
        price:
          type: number
          description: 일반 판매가
          example: 120000
        startPrice:
          type: number
          description: 경매 시작가
          example: 80000
        auctionDurationDays:
          type: integer
          format: int32
          description: 경매 진행 기간(일 단위)
          example: 3
        images:
          type: array
          description: 상품 이미지 목록
          items:
            "$ref": "#/components/schemas/ProductImagePayload"
        location:
          type: string
          description: 거래 위치
          example: 서울 강남구
          maxLength: 100
          minLength: 0
        draftId:
          type: integer
          format: int64
          description: 임시저장 ID
          example: 5
    SaveAuctionDraftResponse:
      type: object
      description: 경매 상품 임시저장 응답
      properties:
        draftId:
          type: integer
          format: int64
          description: 임시저장 ID
          example: 10
        savedAt:
          type: string
          format: date-time
          description: 저장 시각
          example: 2026-04-12 20:00:00.000000000 Z
    UpdateMyProfileRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 30
          minLength: 0
        nickname:
          type: string
          maxLength: 30
          minLength: 2
        bio:
          type: string
          maxLength: 500
          minLength: 0
        profileImageUrl:
          type: string
          maxLength: 500
          minLength: 0
      required:
      - nickname
    MyPageProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nickname:
          type: string
        email:
          type: string
        bio:
          type: string
        profileImageUrl:
          type: string
        location:
          type: string
        verified:
          type: boolean
        rating:
          type: number
          format: double
        warningCount:
          type: integer
          format: int32
        biddingCount:
          type: integer
          format: int32
        sellingCount:
          type: integer
          format: int32
        wishlistCount:
          type: integer
          format: int32
...
    MyLocationResponse:
      type: object
      properties:
        location:
          type: string
        postalCode:
          type: string
        roadAddress:
          type: string
        jibunAddress:
          type: string
        detailAddress:
          type: string
        locationSource:
          type: string
          enum:
          - MANUAL
          - POSTCODE
          - GPS
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT